Check for network reachability
