+++ Sources/Features/Attributes/NetworkReachability.swift
@@ -33,6 +33,11 @@ final class NetworkReachability {
 
 extension NetworkReachability {
     
+    var isReachable: Bool {
+        guard let flags = flags else { return false }
+        return isNetworkReachable(with: flags)
+    }
+    
     var statusName: String {
         guard let flags = flags else { return "unknown" }
         guard isNetworkReachable(with: flags) else { return "notReachable" }
@@ -44,7 +49,7 @@ extension NetworkReachability {
         return "reachableViaEthernetOrWiFi"
     }
     
-    func isNetworkReachable(with flags: SCNetworkReachabilityFlags) -> Bool {
+    private func isNetworkReachable(with flags: SCNetworkReachabilityFlags) -> Bool {
         let isReachable = flags.contains(.reachable)
         let needsConnection = flags.contains(.connectionRequired)
         let canConnectAutomatically = flags.contains(.connectionOnDemand) || flags.contains(.connectionOnTraffic)
